pre walk-forward optimization

+++ core/events/base_event.py
@@ -6,7 +6,7 @@ import uuid
 @dataclass
 class EventMeta:
     id: str = field(default_factory=lambda: str(uuid.uuid4()))
-    timestamp: int = field(default_factory=lambda: datetime.now().timestamp())
+    timestamp: float = field(default_factory=lambda: datetime.now().timestamp())
     priority: int = field(default_factory=lambda: 1)
